Moved data_file and annotations to data directory, fixed import statements
